requirements.txt
# Web Framework
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.26.0

# WebSockets
websockets>=12.0
python-socketio>=5.10.0

# YAML Processing
pyyaml>=6.0
ruamel.yaml>=0.18.0

# Utilities
python-slugify>=8.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# Voice Processing
speechrecognition>=3.10.0
pyttsx3>=2.90

# HTTP Requests
requests>=2.31.0

# JSON Schema validation
jsonschema>=4.20.0

# Logging
structlog>=23.2.0

# Security
cryptography>=41.0.0